pcw8256: Example (in progress) devinput device
- Mice need changing to do their work each vblank and accumulate
- No suport for all up/down events yet
- Still need to work out what mode 1 will mean and define in a globally meaningful way - or add a matches ioctl for say up to 8 'grabbed' keycodes

Hopefully enough of a sketch to start getting joystick support into platforms
at least.

Note: the basic idea of all of this is to bind together all the input events
into a single device so that we don't need all the overhead of poll and big
unixisms. Thus a game can grab the keyboard, and get keyboard/joystick/mouse
events all down one channel

Alan