Library: add the bits needed to build a Rabbit 2000 C library
syscall.s is still a stub until we build the kernel and sort out the syscall
interface a bit more. It might be worth fixing the design flaw in the existing
one for the Rabbit over Z80, and then moving Z80 to it in when we have the
relocatable binary format flag day.