ubee: hard disk and WD1010 floppy support
It probably makes sense to use the block helper layer for this (and we'll want
it for IDE as well). For now though it should work with a .hd3 image.

There is no way (as with the real machine) of easily identifying the physical
disk geometry. We could try requesting sectors in cylinder/head 0 and seeing
where it goes pop.