fweep: squash down to about 18K
There is still a fair bit of fat in the dictionary and tokenization code that
wants sorting out. Also some of the other logic wants tweaking to use pointers
more I suspect.

Once that is done we are nearly ready to add the dynamic buffers so we can page
in and out the game database and try and really run in 8bit micro space