dragon-nx32: Add dual-screen support with CRT9128
Adds CRT9128 card on tty2 with shared keyboard. Switch tty with
ctrl-1 and ctrl-2 (CLEAR-1 and CLEAR-2 on Dragon keyboard).

The 6847 graphics VT interface functions are made private and
these or the CRT9128 ones are called depending on the current TTY.

Add video.h just in case (not needed since multiplexer is asm).

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>