z80: fix a nasty bug with the full bank switching model on swapless
The swap case is ok as it always runs on the swap stack. The non swap case
restores the process kernel stack underneath the curent sp. That in itself is
fine as we have interrupts off. However we then called map_kernel which if
the current sp was above the sp of the save would corrupt the stack.

Instead restore sp first so that we push and pop data below the saved stack
content