dragon-nx32: Only init IDE if we have a known IDE cartridge ROM
Same as we do for SCSI controllers.

Avoids the long detection timeout when there is no IDE controller
present.

Since we are not using the cartridge ROM, it would be strictly
possible to use a controller without ROM but we won't support
that for now. The same goes for unknown ROMs, of course.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>