genie-eg64: New platform (in test only)
The Video Genie EG64 allows you to take a Video Genie (also known as the
Dick Smith System 80, the TRZ80 and a few other names) with either of the
expansion units and the EG64 fitted up to 96K. It's intended for running CP/M
but it's just about big enough to run Fuzix badly providing you have a swap
device on hard disk.

Another goal of this port is to figure out how to move a lot of TRS80 driver
code into dev/trs80 to share between these two ports, a Genie III port maybe,
and also potentially things like the Lobo-Max, Guepard., SysData III and so on.