sam: further debugging work
We now get to the bootdev prompt with interrupts running. No keyboard input
yet but that's next to investigate.

Note that there's a stop after loading so if you are using simcoupe you need
to hit shift-F10, ctrl-downarrow escape to set it loose (or set up debugger
stuff first as needed)