utils: teach fsck mount and remount about the root name
This allows us to stop all the substroot stuff as with fsck split we now have
enough space. This actually turns the whole thing into a win. We trade the
extra exec's of substroot for a fork/exec of fsck.

Not only that but with a bit of thought we can later optimise the single
case fsck to kill the fork off