sbcv2: Get rid of the boot via CP/M hack and make it bootable
This needs a little care because it's not well documented and also

1. If you tell ROMWBW your loader is from F200 to F3FF it loads all of memory
and goes boom. Tell it F400 and it's ok

2. ROMWBW lacks the notion of an 'I/O bank' so you can't do disk I/O into the
low 32K directly.

That aside ROMWBW has a nice easy interface so the boot loader is quite trivial.
Not quite as trivial as the Microbee but not far off!