New MOOH platform with 8K flexible banks
This is in good enough shape to try out levee :)

Adjusting the program allocation sizes with chmem seems to work fine.

A few TODOs:
The virtual consoles currently overlap with the graphics framebuffer so
using them doesn't look pretty. Trials at moving them fails badly, maybe
related to next issue:

We need to disable interrupts in the code accessing video memory,
because common is not there then. Or deal with this in the interrupt
handlers. Trials at disabling interrupts have not been successful yet.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>