tools: add relocatable binary support
We can either fight the tools and rework the linker or follow the long standing
tradition of linking it twice 256 bytes apart and doing it that way.

For now lets do it the tool independent way so people don't need a customised
sdcc chain.

The fcc part is done, the reloc tool doesn't most of the work but it doesn't
yet write the relocs into the binary and we don't have reloc awareness in the
loader.