msx1: revisit this platform
Now we have the infrastructure and simpler examples of ROM based kernel with
a RAM user mapping see if we can get MSX1 64K + kernel in cartridge to work
using a swap device.

For mappers it doesn't really make sense to keep them in this port. An MSX1
with MSX2 mapper would best be handled by fixing the MSX2 port to support all
the different VDP types. MegaRAM is probably its own port but one to tackle
once this works and also once we have relocatable binaries.