lowlevel-z80: switch our handling of interrupts
This has implications for all platforms and breaks everything but it's worth
the pain to avoid dealing with the NMOS crawly interrupt horrors.

While breaking everything also introduce _di versions of the map functions
to allow problematic platforms to shortcut having to IRQ disable in these
cases (which are also the fast paths)

As we always save and map kernel fold those into one to optimize the IRQ
path a shade.