kernel: allow for cases with only a small window in the base page
If you have a 4x16K or similar mapping model and most of the upper page
is taken up by discard and common then there may not be 4K to randomly play with.

In that case firstly we overflow and miscalculate the pages, secondly we then
scribble on stuff we shouldn't.

Instead tighten it up so there needs to be only 512 bytes (plus vectors etc
- 256 bytes on Z80) of space.