kernel: rework usage timers
The old UZI code we started from defined the various time fields in both proc
and udata but used only the udata one. Perhaps an unfinished attempt to move
from proc to udata. While it saves memory keeping it in udata (20 bytes a
process or 320 bytes) we were wasting that memory anyway.

Go back to keeping it in proc as that avoids some messiness and standards
violation in the code, and makes ps work nicely. Possibly we can later move
cutime and cstime into udata but that needs review (otoh it would leave us
80 bytes up on where we started)