zxdev: second attempt at exploring a spectrum port
This time assume we are going to work with interfaces that can put us in the
low 16K. That is true for a lot of disk interfaces and also for many of the
Russian clones so they could run CP/M.

That gives us a much saner memory map to work from.

We get part way into boot - need to fix up the memory mappings and udata
to make further progress, then debug the DivIDE disk interface logic.

The loader works though