zxdiv: Fix various things
Map the code low
Mark code as code1 so it banks right
Fix the writable code references
Sort out the udata mapping
Fix the __memcpy calls to call our bank code not invalid sdcc lib code
Remove strlcpy we don't use
Fix minor fork bugs

With this we get to the point of init running and forking. No further but
we really need swap on to make any real progress.