zx128: Move buffers to 0x6000 and dynamically reclaim space
This gets us dynamic buffers as we want and also frees up the space between
0x4000 and 0x5FFF so that we can use the ZX Spectrum classic screen mapping
and leaves a bit of space in case we need to use any ROM hooks that want to
use system variables and BASIC hooks.

We are a bit tight between 0x2000 and 0x3FFF but have enough space for now.

Freeing up the space means we can next move the screen down. That makes the
screen potentially directly addressible from user space which is a huge win.

It also means we'll have a load of free space in CODE3 which is what we want
as well as probably room to migrate the font about a bit