cromemco: further updates - gets us to the boot prompt
The tty interfacing via IM2 mode is a bit hackish for now. Unfortunately because
the emulation of the tu-art in Z80pack is currently extremely buggy we are
basically forced to use IM2.

Eventually we want to do IM2 properly anyway. The entire system is built around
a design that has proper interrupt handling.

With these changes sorted interrupts now work, and you can fail to find a root
file system. Next task is getting the fdc and root fs working.