process: Fix nready panic and hangs when using rts/cts
Actually what happens is that spend long enough waiting to write the byte
that we hit the poll timeout. In the usual case it doesn't matter but if

1. You hit a key
2. We are idle
3. We are doing idle key polling
4. The echo is blocked for a long long time

We then try and sleep the current process, only we are in idle so there isn't
one. Idle is really like being in an interrupt.

Now arguably we should fix tty output to retry the echo write until it gets
kicked for timing offences and then give up but tht means auditing all the
ininterrupt uses and also encourages the existing bad behaviour of not running
a small tx buffer.

So if it annoys you fix your tty output code 8)