zxdiv48: Initial code for 48K ZX Spectrum with DivIDE plus or ZXCF
Queueing this up for weekend testing.

It seemed a shame that the classic squishy key machines were being left out
so there was only one thing left to do - fix it.

It's a curious combination of the 128K/DivIDE model and the single process
in memory model. The kernel is banked four ways in the low 16K using memory
from the DivIDE Plus or ZXCF card. The area between the screen and 0x8000 just
about fits enough data for a minimal system providing the disk buffers are
banked. From 0x800-0xFBFF is then used for the application space with FC00-FFFF
holding the stacks, udata and a 256 byte space we may yet need.

The rest of the 512/1024K of RAM on the interface is then used as swap.

Task switching is slow (two x 32K ldir) but that is no worse than the disk
load/save used by other single process in memory platforms.

The biggest problem is the limited amount of unbanked kernel RAM. With the
screen at 0x4000-0x5AFF we only have a shade over 8K for all our common,
our data, constants and bank switching stubs. To get it to fit we end up
limiting our user space to 31K (versus 31.5K on most other 32K switched bank
platforms).