pentagon: pentagon1024
Copy the work so far over to a new platform because the older Pentagon systems
don't have the RAM0 low mapping and will need something a bit different.

As we are dependent upon the 1024K features we can just hardcode 1MB of RAM.
Not that you can use it all right now. 16 processes at a time is 512K and given
the task switch cost it's not clear how you'd use 32 processes in memory sanely.

We could add a 512K RAM disc but then with an vaguely modern CF card the
difference between banked RAM copies and CF card I/O is actually in favour
of using the CF card.

Maybe we'll just have to invent banked applications for such systems one day.