syscall_exec16: Fix BSS wrap bug
If the binary fits but it has a huge BSS value then kernel wrongly allows it to
start. Because of some other bugs and quirks it ends up running sanely for most
platforms but the BSS is not cleared.

This breaks the current builds of /bin/sh so you'll need to build a new one with
the tools patches that follow this applied.