usermem: correct wrapped sizes
There are cases it is valid to pass a base and size that wrap, if your user space
runs in the high memory area. On the Linc80 this shows up and breaks the
n_open optimizations because instead of FE00 200 giving 100 as it should it gives
an error.

This was also the bug that happened to save us with binaries with the wrong BSS
size