z80: brk to the correct point in the relocatable binaries
In the cases where we couldn't put the relocations into the end of data and
make it BSS we end up with a binary that is a bit bigger than expected. Most
stuff doesn't care but for correctness and tight memory cases use brk() to
put the end of BSS back at the right point so future sbrk() calls give the
expected result.