kernel: change p_time (the start time)
We really want to know this in ticks not time otherwise it gets nonsensical
when you set the clock. In particular init is almost certainly created with a
bogus time of day.

It's then easy enough to fix ps to instead ask for the tick counter and do
the maths relative to current time to work out what is going on.

Also conveniently fixes the 2038 issue