mtx: Memotech updates
Start putting in place the stuff for the CFII and the Propeller based 80 column
card. Also fill in some of the graphics and other bits. Fix the maximum memory
banks.

There are still some really ugly things to deal with. Do we expose the propeller
directly or via the VT layer. How do we align that with the graphics ioctls.

Stil need to think about UDGs