sunrise: redo the bounce buffer logic
We can't just use tmpbuf because there might not be a free buffer and we'd
then get into a mess recursing through our I/O subystem. Instead grab a bit
of spare high memory for the moment.

Once the buffer reclaim logic is added we will turn discard and friends into
buffers but keep the last 512 bytes before FFFF as a bounce buffer that drivers
can share.

(FFFF itself is slot switching so magic and reserved)