z80-bios: development tree for a BIOS based Z80 Fuzix
This is going to be needed for things like S100 where your machine is
defined by a BIOS layer not hardware. We can't unfortunately use the CP/M
BIOSen for this:

CP/M 2.2 has no understanding of banking and no way to handle output blocking
CP/M 3 banked has some of the banking and blocking covered but has no
way to hook interrupts or events
MP/M is closer but has a whole pile of assumptions that look hard to deal with
plus very few machines have an MP/M BIOS anyway.

So start with a clean design (as if firmware is ever clean).

Alan