devide: Fix map_buffers/map_kernel mismatch
This corrects the fact that map_buffers is not mirrored with a map_kernel.

Platforms that null map_buffers and expect map_kernel not to be called are
going to need fixing from this change but there isn't any way to put it off
because there are platforms that need the correct behaviour to work.