kernel: Fix the timer/alarm queue logic
Rely upon the flag to stop duplicate insertion (if we are the last entry and
inserted then it's still NULL next pointers)

We can exit() while in the queue. Fortunately we already walk the process table
for re-parenting so we can fix the list up for free at the same time.