ubee: Fix memory banks
The original code wasn't correctly adding all the banks. Fixing that then
broke everything because those missed adds also hid the fact that we should
have been avoiding page 2 not page 0. Physically the top 32K is bank 0 block 0
but logically bit 1 is inverted when ROM is mapped out so we were omitting
the wrong page.