micro80: initial port for Bill Shen's latest design
Z84C15 (single chip Z80 and peripherals) with 128K RAM and a CF adapter.

It's only got 128K of RAM of which we can use 120K but otherwise makes a
near perfect Fuzix box. Because of the limits on the memory mappings the
common is placed at 0000-0FFF. This does mean we can't run CP/M emulation so
it may be worth revisiting.