rc2014: Turn on DMA for fork()
There are other potential uses but the user memory one requires that we fix
the user copiers and the CF one requires writing some DMA CF driver code.

Also the CF one will probably need the IDE driver layer optimizing or we are
just going to walk the bottleneck up the stack.