rc2014: Speed up copies to and from user space.
Because the area between 0x4000 and 0xBFFF contains only code we know that
no user copies will have those kernel addresses (and indeed as this area
is banked it would a disaster if they did).

This means that for any given user address we can map the containing 16K
bank at 0x4000 and the next one at 0x8000 and be sure we can copy up to 16K
in one go and we can loop for the very rare cases we do.

Because the memory is mapped we can just ldir instead of keep bank flipping
and this in turn makes the copies *much* faster.