rc2014: QUART support and TMS9918A fixes
With these the TMS9918A seems to work right as a clock and the QUART
also appears to be working although all the port speed combinations could do
with some testing and cleaning up.

Might also be worth seeing if we can use the per channel BRG select extend
stuff.