syscall: First phase of fixing the Z80 syscall interface and converging with 8080
Stop saving DE, start saving BC
Expect the arguments 2 bytes offset, thereby avoiding all the crap in the
syscall.s code.

This converges the syscall side of the interface but we still need to deal with
the signal differences. After that we need to do the general switch of the
8bit platforms to the new headers and VDSO like block.

In other words this ABI is temporary so don't code to it!

This patch set will require compiling a whole new userspace for Z80. You
probably want to wait until everything is done before working on Z80 with
this lot.