libs: Update all the Z80 crt0 code for the signal handler changes
With these in place the same binary ABI is present for 8080, 8085 and Z80 which
means we can run 8080 binaries on all three. This isn't the end of the binary
format changes. The next step to be done is to change the FZX1 headers to a new
FZX2 format - so this is still a temporary ABI right now