6303: get the kernel to the point it compiles and links
There are still routines to be written (see tricks.S), and a lot of work to
do with the C compiler turning it from 'writes code for' to 'writes working
code for'.

However getting this far has been a good way to stress test the front end,
the assembler and the linker.