rc2014: Move the primary 16x50 probe to 0xA0, fix up SIO vectors
This last one is a bit obscure. If you have multiple IM2 capable devices in
IM1 mode then they will all see the IORQ+M1 cycle and try and put a vector
on the bus at once. That is fine providing they all put the *same* vector on
the bus. Whilst the CPU igores it we don't want the cards fighting.