68hc11: add the first pieces of 68HC11 platform support
There is a lot left to do here but it makes sense to begin building this
in parallel with the 6303 work. There are not many differences except toolchain
ones. The 68HC11 has another register, which changes stuff like SWI offsets,
but otherwise it's mostly toolchain differences (_ on names etc, and globl v
export etc)