vile: fix failures on machines with a larger TPA
vile works fine if you have a TPA of about 48K or less but on SC126 and some
other platforms it succeeds in allocating over 32K for the buffer, at which
point the indexing breaks due to overflow.

The indexes can be unsigned, and also sized according to the platform. This
also fixes it breaking on 68K with larger files.