rc2014-68008: First pieces for 68K RC2014 support
This will boot into the early stages of the kernel if you are using the current
ROM image. It will not get further as ROM and kernel changes are needed to
handle the ROM based exception vectors. The ROM needs revector support adding
and the kernel needs to know how to use that.