malloc: fix the kernel malloc
Found and reported by @pawso-arm

This mostly worked by luck splitting the last block each time. With this change
we should be doing best fit.

This won't fix at least one of the underlying problems with flat model right
now on some systems because we've got udata blocks getting allocated all
over the map and pinned forever so they break the map up badly.

Preallocating them as rc2014-68008 does works but the right answer is
to free them as they become dead. That's tricky because on the exit
to switchin path they don't actually die until we've switched because
they contain our stack. May need to free them on the path out of switching
if there is one left hanging around (and we can free others safely in exit)