rc2014-68hc11: turn on a lot of the interrupt stuff
We now have timers, we have serial interrupts and pre-emption. Most of it
actually works. At this point however asynchronous serial transmit gets
confused somewhere so serial tx is still synchronous for now.

Pre-emption crashes the system if we pre-empt in the timer event. This needs
tracing down. Timers in general however work if pre-emption is only enabled
for system call level checks and dates and clock progress all appear to be
working.