amstradnc: documentation update
At this point we need to debug the PCMCIA rd driver and update it both for the
newer I/O behaviour and possibly for NC200 swap (or just steal some of it as
RAM based on some kind of signature checks).

At this point we crash running init because only the low 16K of init ended
up in the right place.