z80: the thunked case is special so we need to fix it
In the thunked case we need jump through the platform stubs. This wasn't
apparent as a bug until the binaries were taught to use the stubs, as the
old RST entry point is correct.

This will need a tiny tweak to each thunked platform but what probably wants
to happen instead is to use macros for the inline bank flips and have a
generic template because for a straight (non 32K split) system the code is
basically identical except for this on each system.