amstradnc: Fix the NC100 suspend resume
The NMI auto power off now properly saves and restores the state so that you
can wake the machine and alt-X to continue your session. It also means we now
correctly handle OS switching so you can suspend Fuzix, run NC100 apps, and
then return to Fuzix. In fact as the filesystem and OS live on the same card
you can actually suspend, and put the card into another NC100 and continue.