doexec: clean up passing the base address
We need to be able to pass a base address that is not the default load address
because there will be cases we want to relocate a program to different base
addresses depending upon type or memory mappings.

This is the first half of the story - the signal helper is a much uglier
problem.