dk-tm4c129x: initial pieces for the port
This is taken from the work that was

Signed-off-by: Paul Osmialowski <pawel.osmialowski@arm.com>

and reworked to avoid needing to pull in all the crap SDK stuff that hides
everything in a giant pile of macros.

Probably I've added some bugs, the clock/pll bit in particularly is a shade
hairy and I don't have any way to test this stuff.