6502: Implement relocatable binaries
The 6502 doesn't seem to have any way we can do this in user space so we do
the work with a helper stub as we execve(). This creates one minor oddity in
that binaries start by using brk() to be sure their size is right because
they know about relocation padding and the relocator doesn't know the true
size.