kernel: turn off the other load address hackery
We have relocations where it matters, except 6809 which can be fixed easily
enough and a few platforms where relocation is expensive and fairly pointless
like 8080/8085 which always load at 0x100 and seem unlikely to ever need
relocatable binaries. except to run on Z80 systems. 8080 reloc isn't hard to
do anyway if we need it.

Removing this ability makes a ton of ugly bits of asm magic a lot easier