time: Switch all the time code to just use the struct
We try and be clever about uint64_t v 2 x uint32_t across the syscall interface
but it just causes alignment and other surprises. Simply pass the time across
as the struct always, and keep the ordering consistent too.

This breaks the ABI on the big endian boxes but none of them have a stable API
yet anyway so it doesn't actually matter.