networking: Lots of bug fixing on the code rewrite
This takes us to the point that you can use the networking stack again and
the basics for the Wiznet work. This isn't complete yet there are two things
to resolve of note

1. The WizNET state machine in the driver and core code doesn't correctly
work out how to use and reuse sockets if the remote closes first in particular

2. The current wakeup mechanism is a hack and racy. Fixing this needs some
thought about how a proper wakeup/sleep model will work whilst keeping the
kernel and network split.

The W5100 code itself is a lot more stable than the old one thanks to the IRQ
blocking fixes. The W5100 right now also does not support

- Correct behaviour when you bind a UDP socket (needs software address filter)
- Incoming connections - this is messy
- Loopback - probably won't fix