w5300: more debugging and fixes
This gets us to the point the addresses used are right, things connect and
disconnect and you can read and write data. The write side however still
isn't quite right - some bits are going missing, and when you hit the end
of the buffer having received some data we get confused.