esp8266: add the packed and other defines we need, move to esp8266
The current build is terminally confused about whether its an lx106 build or
an esp8266 build. Given the esp8266 is basically unique because of the way
the tensilican core is generated, and the fact there is no other tensilica
core we care about, and the next generation ESP has gone RiscV anyway.

ESP32 is a different tensilica set up and we'll figure out later if anyone
ever care.