esp8266: first pieces of signal handling, and int/syscall tracking
The tracking is needed by the switching code so that it knows when
it can pre-empt and the correct interrupt state to set when the
task resumes in switchin.

The signal handling covers the easy cases for the moment, so a
signal is delivered on syscall exit and synchronously on an
exception but not an asynchronous signal on an interrupt.