esp8266: go multitasking
The changes are actually pretty small. The switchout logic now has to swap the old task back out (this
can only happen in fork when single tasking). The switchin logic must restore the correct interrupt state.

We turn the feature on and we also force a relatively slow task switch rate as we don't want to thrash
if two things are running. In normality with a single user we actually only have one running task, and
we have almost no false wakeups as the entire kernel is designed to avoid them.

At the moment we are doing the simpler 'swap parent' approach. For better performance we should swap the
child so the parent runs to wait() then the child runs as a general case. At the speeds of I/O we have
this isn't a big deal.