Library: add an initial tool for generating shared library structures
The idea is to generate code with this and then link the resulting file with the library. That gets you
a jump table headed versioned library which can then be mapped appropriately for systems where it will
be useful (eg those with a large flat addresss space, lots of flash and little RAM).

In the simple form this is quite limited because we can't include functions that have any static data
or functions that call functions which do. However for the simpler setups like the ESP8266 where the
user data is at a fixed location we can just reserve a bit of memory for the libc static data that
we care about (mostly errno).

This is still not enough to deal with functions that use static variables resident in the application
and a few support ones can move (environ for example), but some will just have to stay out of the
shared library - likewise those with big static objects.

Quite a few of our candidates are not going to be hurt by these restrictions. The 6502 boxes in
particular have a big common cc65 compiler support library that user and kernel need and it uses very
little static data and hasn't really changed in years.