banksplit: model for systems with a small switable common
There are a bunch of systems where you can divide memory so you have a large program area that is banked
and also a high space which can be separately flipped as opposed to the more usual fixed common.

Having this flippable common space really speeds Fuzix up but the platforms that use it don't always allow,
or easily allow, the memory space that is in the common area to be mapped anywhere else. This makes swapping
the usual way very tricky. Instead what we do is let the platform provide a small asm wrapper to call the
swapout() method temporarily running in the context of the victim being swapped, and on a private stack.

At the moment this is most relevant for Z180, but in theory could also be applied the RC2014 MMU and
8085/MMU cards that provide a 56K/8K split in the same style.