tms9995: walk over the low level code
This is complex and will take some debugging or maybe a rethink. The need
to use contexts for interrupt speed, and the fact the processor can only
restore the status as part of a workspace makes it more fun than it needs to
be.

On a traditional TI/990 or a bigger micro with 16bit RAM it would be easy -
each process would have a workspace as would each interrupt. Then we'd just
have to do some basic push/pops of contexts. However we've only got 256 bytes
of fast context memory for registers and that's simply not enough to work
in that style.