6800: More usermem and support fixes
This gets us to the point we load a binary, set everything up and then jump to
it, at which point we find garbage at those addresses.

Need to look further at uput and at buffered disk I/O paths because the
junk shows in 0x01xx which ought to be read through the cache as we read the
header first.