kernel: fix up the other bug from the conversion
With this in place we should be able to fairly easily do other file system
sizes. The current assumptions built in are

1. No inode crosses a block boundary
2. No directory entry crosses a block boundary

The former is encapsulated in the blknnn routines so a platform can implement
the needed padding. The directory crossing assumption is in the core code.

For sane block sizes none of this should matter. For insane block sizes like
the Centurion it might