Add files via upload
Modifications in config.h and devsdspi.c for using the SD-Card on the Cytron "Maker Pi Pico" for swap and root. These are only part-integrated with the cmake build file, in principle they also make sure that the "Debug" build is actually debuggable by not moving it from Flash to RAM.

There's probably not many people with this type of board, but these highlight the changes necessary to use an SD-Card with arbitrary connections rather than the Pico's own memory.