n_open: Fix a bug and a bit of daft design
An out of bounds address now correctly returns the error code.

In addition there was a bug in unlink because parent was uninitialized and
n_open assumes the caller set it to NULLINODE if using it. So this doesn't
happen again force it to NULLINODE initially in n_open instead.