kmod: initial pieces to start support for kernel module loading
Record the area beyond the required buffers, and allow it to be reclaimed
from the top down for loadable modules. This provides us with as framework
to load stuff over discard space when we would otherwise have no room, or
in future for platforms to run with banked loadable drivers.

At the moment all the smarts are in the loader. This may want revisiting.