rbc-mini68k: Mini68K merge
This port is provided by John Coffman. It's been modified a bit since just
to make it fit the tree better an fix a couple of things.

There is work to do in a few spots but this works OK on the emulator so
hopefully works on the real board too.