rbc-mini68k: rewrite the tty code
The modular part is based on the RC2014 code and the 16x50 code is a refinement
of the code used in a couple of other places. This also hooks into the new
extensions for the tty layer to support all the error and break reporting.

We still need to sort out the console and pick up the BIOS baud rate by
scanning the speed table.