ns32k: file relocations
The ns32k tools don't include relocations even if asked to, so instead do it
the old fashioned way. Link twice and compare. We output an a.out file with
relocations in our binfmt_flat format. This is a temporary bodge.

The plan is to switch to aout with our reloc format on 68K as well, and move
the binary 32bit format to a.out like a real grown up computer.

Hopefully this works out and there are no strange relocations going on other
than straight 32bit addresses to mess up the plan.