2063: Another retro Z80 board
Z80, CTC, SIO and bit bang SD card. Somewhat similar to the Z80Retro
in that respect but memory banking is the rather simpler fixed 32K
banked low 32K arrangement and the SD is done differently

This gets us up and going - needs to switch to IM2 and sort out
various odds and ends.