rcbus-ns32k: turn off split I/D
The compiler chain has a devout belief that variables are to be accessed
via xx(pc), which is actually understandable given how things work but does
mean code and data need to live together.

In theory we could modify flat to support some kind of "partially copy" the
single segment hack on fork etc but that's ugly so we'll skip it for now.