flat_small; paging test case
Not yet complete. For one the queued swaps and the non queued page out
don't combine to give a correct result. In particular we will try and
swap out the oldest pages which means we'll keep trying to swap them
all with the same page and that would have to happen with the swaps
not queued relative to them.

Possibly the queue trick isn't worth doing. Alternatively we might need
to simply keep a list of pages by number to swap out and for each swap
spot if they are on the list and if so do a swap out and copy not an
exchange.