flat_small: get the first paging manager working
This is designed to cover systems where there is enough RAM to do useful
work but where swapping is expensive or needs to be kept to a minimum for
other reasons. On a system with a fast hard disk interface and low memory
the straight forward "one process in memory" model is likely to be faster
and smaller.