8086: Application makefiles
Nice and simple now we've got the big cleanup mostly done.

This gets us to the point we build all the apps and produce binaries
with a wrong crt0 in the wrong format.

Needs the toolchain from https://github.com/tkchia as the older toolchain
will ICE trying to build dc