zxuno: rework to be simpler and get it part working
The previous clever plans don't work out as the Uno has some interesting
quirks and bugs that screw it up totally.

Instead use the upper 48K of the dock and exrom for user processes, use the
128K mapping for main RAM and force map and pin the DIVMMC in the low 16K
to avoid the DIVMMC paging bugs.

This costs us a chunk of memory and other things but it doesn't seem possible
to avoid it given the breakage in the system.

At this point we get to fork then it messes up