rcbus-z180: SPI improvements - speed switching, allow "wiznet" option
On a non SC126 style system it is possible to wire \CS low and use the CSIO
plus a reset line for a Wiznet. Allow the user to specify a "wiznet" boot
option for this.

Also allow for the case where we have a wiznet and an ancient slow SD card
on the same bus and switch speeds as we go. This also fixes the fact that
if only a wiznet is probed for nothing set up the CSIO clock